Add files via upload
Mainly to test an idea. Feel free to edit xdat.txt and ydat.txt as you see fit.
Although it might be best to have them in the form of one CSV, for readability sake I want to have the separated. 
Else, could store as "data.txt" and rewrite the lines for x and y as
    data = load('filename.txt')
    x = data(:, 1);
    y = data(:, 2);
similar to the lines in test.m.